Makes get_nonroot_uuid a standalone util
